+++ components/download/public/common/download_utils.h
@@ -6,6 +6,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef COMPONENTS_DOWNLOAD_PUBLIC_COMMON_DOWNLOAD_UTILS_H_
 #define COMPONENTS_DOWNLOAD_PUBLIC_COMMON_DOWNLOAD_UTILS_H_
 
+#include <stddef.h>
+
 #include <memory>
 #include <optional>
 
@@ -154,7 +156,7 @@ COMPONENTS_DOWNLOAD_EXPORT base::TimeDelta GetExpiredDownloadDeleteTime();
 COMPONENTS_DOWNLOAD_EXPORT base::TimeDelta GetOverwrittenDownloadDeleteTime();
 
 // Returns the size of the file buffer that reads data from the data pipe.
-COMPONENTS_DOWNLOAD_EXPORT int GetDownloadFileBufferSize();
+COMPONENTS_DOWNLOAD_EXPORT size_t GetDownloadFileBufferSize();
 
 // Utility function to determine whether an interrupted download should be
 // auto-resumable.
